requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
cloudscraper==1.2.71
beautifulsoup4==4.12.3
lxml==5.3.0
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.3
Pillow==10.4.0
